# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MagFin API Documentation",
        "description": "MagFin API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "[development] MagFin App Server"
        }
    ],
    "paths": {
...
        "/api/super-admin": {
            "get": {
                "tags": [
                    "Super Admin"
                ],
                "summary": "Get all data for super admin",
                "description": "Get all data for super admin",
                "operationId": "getAllDataSuperAdmin",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "super_admin": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SuperAdmin"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "full_name": {
                        "type": "string"
                    },
                    "shop_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "balance": {
                        "type": "integer"
                    },
                    "role": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "trahs_manager_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "is_verified": {
                        "type": "boolean"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TrashManager": {
                "title": "TrashManager",
                "description": "TrashManager model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nama_pengelola": {
                        "type": "string"
                    },
                    "tempat": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "super_admin_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "users": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object"
            },
...
            "SuperAdmin": {
                "title": "SuperAdmin",
                "description": "SuperAdmin model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nama_pengelola": {
                        "type": "string"
                    },
                    "tempat": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "trash_managers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TrashManager"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}